remove copying again
Copying the typing context and associated _globals turned out to be a
futile endeavour, so instead we now set and unset the variable before
and after compilation.
