+++ numba/typing/typeof.py
@@ -169,7 +169,8 @@ def _typeof_list(val, c):
         raise ValueError(
             "Cannot type list element of {!r}".format(type(val[0])),
             )
-    return types.ListType(ty, mutable=False)
+    #return types.ListType(ty, mutable=False)
+    return types.List(ty, reflected=True)
 
 @typeof_impl.register(set)
 def _typeof_set(val, c):
